Widen QSOP modulus to uint64_t, add cry gate, take a calloc-avoidance perf win
qasm2sop --approx previously refused any circuit whose error budget needed a
modulus beyond uint32_t ("requires a modulus larger than 4294967295"), even
though the single-Fourier-mode solver added earlier is already O(1) in r.
Widen qsop_instance_t.r/.constant/.unary (and the matching qasm2sop pipeline)
to uint64_t so qasm2sop can actually produce what the solver can already
consume. Old count-table/all-modes-Fourier backends (components, bruteforce,
branch, treewidth, rankwidth, qsop_wmc's residue/residue-fourier encodings)
now cleanly refuse r > 2^32-1 instead of silently truncating, since widening
their O(r)/O(r^2) allocations wouldn't help them handle a huge r anyway.

Also: add the cry gate to qasm2sop (same H/phase conjugation apply_ry already
applies to apply_rz, applied to crz instead), and switch the complex-mode
factor allocators from calloc to malloc now that every caller either fully
overwrites its buffer or explicitly zeroes only where it accumulates
(measured via callgrind: memset share 21.2% -> 12.0% on a profiled instance).

+++ include/dlx4sop/qsop.h
@@ -7,12 +7,12 @@
 #include <stdio.h>
 
 typedef struct qsop_instance {
-  uint32_t r;
+  uint64_t r;
   uint32_t nvars;
   uint64_t norm_h;
-  uint32_t constant;
+  uint64_t constant;
 
-  uint32_t *unary;
+  uint64_t *unary;
   uint32_t nedges;
   uint32_t *edge_u;
   uint32_t *edge_v;
